requirements.txt
pydantic
sqlite-utils
numpy
faiss-cpu
sentence-transformers
